+++ src/main/java/com/imcode/imcms/mapping/orm/FileReference.java
@@ -19,7 +19,7 @@ public class FileReference {
 	private Long id;
 	
 	@Column(name="meta_id")
-	private Integer metaId;	
+	private Long metaId;	
 	
 	@Column(name="filename")
 	private String filename;
@@ -76,11 +76,11 @@ public class FileReference {
 		this.id = id;
 	}
 
-	public Integer getMetaId() {
+	public Long getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Integer metaId) {
+	public void setMetaId(Long metaId) {
 		this.metaId = metaId;
 	}
