+++ src/main/java/com/imcode/imcms/mapping/orm/HtmlReference.java
@@ -10,7 +10,7 @@ import javax.persistence.Table;
 public class HtmlReference {
 	
 	@Id @Column(name="meta_id")
-	private Integer metaId;	
+	private Long metaId;	
 
 	@Column(name="frame_set")
 	private String html;
@@ -23,11 +23,11 @@ public class HtmlReference {
 		this.html = text;
 	}
 
-	public Integer getMetaId() {
+	public Long getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Integer metaId) {
+	public void setMetaId(Long metaId) {
 		this.metaId = metaId;
 	}	
 }
