+++ src/main/java/imcode/server/document/textdocument/TextDomainObject.java
@@ -62,7 +62,7 @@ public class TextDomainObject implements Serializable, Cloneable {
 	private Long id;
 	
 	@Column(name="meta_id")
-	private Integer metaId;	
+	private Long metaId;	
 	
     /**
      * Text filed index in the document.
@@ -214,11 +214,11 @@ public class TextDomainObject implements Serializable, Cloneable {
 		this.id = id;
 	}
 
-	public Integer getMetaId() {
+	public Long getMetaId() {
 		return metaId;
 	}
 
-	public void setMetaId(Integer metaId) {
+	public void setMetaId(Long metaId) {
 		this.metaId = metaId;
 	}
